Wed, May 3, 2023, 3:48 PM +00:00

+++ Psets/pset1/interpreter.py
@@ -11,16 +11,16 @@ For instance, if the user inputs 1 + 1, your program should output 2.0. Assume t
 
 expression = input("Enter your expression:  ")
 
-x, y, z = int(expression.split(" "))#Check
+x, y, z = expression.split(" ") #Check
 
 match str(y):
     case "+":
-        print(x + z)
+        print(int(x + z))
     case "-":
-        print(x - z)
+        print(int(x - z))
     case "*":
-        print(x * z)
+        print(int(x * z))
     case "/":
-        print(x / z)
+        print(int(x / z))
     case _ :
         print("Enter valid math expressions!")
